pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "gcp-sales-pipeline"
version = "0.1.0"
description = "Pipeline to join BigQuery product data and GCS sales data."
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
  "pandas>=2.2.0",
  "google-cloud-bigquery>=3.20.0",
  "google-cloud-storage>=2.13.0",
  "pyarrow>=15.0.0",
  "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "pytest-mock>=3.12.0",
  "black>=24.0.0",
  "flake8>=7.0.0",
  "pre-commit>=3.7.0",
]

[project.scripts]
gcp-sales-pipeline = "gcp_sales_pipeline.cli:main"

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503"]

